hot fix ,prevent overflow

+++ evergreen_server/Item.h
@@ -11,6 +11,6 @@ public:
 	virtual bool UseItem(ContentsEntity* const owner)noexcept { return true; }
 public:
 	int8_t m_itemDetailType = -1;
-	int8_t m_numOfItemStack = -1;
+	int64_t m_numOfItemStack = -1;
 };
